Extract zerver/lib/fix_unreads.py.
This is a pure code move.
